2.0.3 is release

+++ src/main/java/com/dgsoft/erp/business/order/OrderCancel.java
@@ -40,14 +40,14 @@ public class OrderCancel {
     public void cancelOrderMoney() {
         log.debug("cancelOrderMoney OrderId:" + orderHome);
         orderHome.getInstance().setCanceled(true);
-        if ("persisted".equals(orderHome.persist())) {
+        if ("updated".equals(orderHome.update())) {
             initProcessInstanceHome();
             processInstanceHome.suspend();
             businessCreate.create();
         }
     }
 
-    private void initProcessInstanceHome(){
+    private void initProcessInstanceHome() {
         processInstanceHome.setProcessDefineName("order");
         processInstanceHome.setProcessKey(orderHome.getInstance().getId());
     }
@@ -57,13 +57,15 @@ public class OrderCancel {
         BigDecimal operMoney = BigDecimal.ZERO;
         for (AccountOper ao : orderHome.getInstance().getAccountOpers()) {
             if (ao.getOperType().isAdd()) {
-                operMoney.subtract(ao.getOperMoney());
+                operMoney = operMoney.subtract(ao.getOperMoney());
             } else {
-                operMoney.add(ao.getOperMoney());
+                operMoney = operMoney.add(ao.getOperMoney());
             }
         }
+        log.debug("backMoney:" + operMoney);
         if (operMoney.compareTo(BigDecimal.ZERO) != 0) {
-            orderHome.getInstance().getCustomer().getBalance().add(operMoney);
+            orderHome.getInstance().getCustomer().setBalance(
+                    orderHome.getInstance().getCustomer().getBalance().add(operMoney));
         }
         orderHome.getInstance().getCustomer().getAccountOpers().removeAll(orderHome.getInstance().getAccountOpers());
         orderHome.getInstance().getAccountOpers().clear();
@@ -73,17 +75,20 @@ public class OrderCancel {
 
         for (NeedRes needRes : orderHome.getInstance().getNeedReses()) {
             for (Dispatch dispatch : needRes.getDispatches()) {
-                for (StockChangeItem item : dispatch.getStockChange().getStockChangeItems()) {
-                    item.getStock().setCount(item.getStock().getCount().add(item.getCount()));
+                if (dispatch.getStockChange() != null) {
+                    for (StockChangeItem item : dispatch.getStockChange().getStockChangeItems()) {
+                        item.getStock().setCount(item.getStock().getCount().add(item.getCount()));
+                    }
+                    orderHome.getEntityManager().remove(dispatch.getStockChange());
                 }
-                orderHome.getEntityManager().remove(dispatch.getStockChange());
             }
             needRes.getDispatches().clear();
-            needRes.setStatus(NeedRes.NeedResStatus.CREATED);
+            needRes.setStatus(NeedRes.NeedResStatus.REMOVED);
         }
 
-        for(OrderItem orderItem: orderHome.getOrderItemByStatus(EnumSet.allOf(OrderItem.OrderItemStatus.class))){
-            orderItem.setStatus(OrderItem.OrderItemStatus.CREATED);
+        for (OrderItem orderItem : orderHome.getOrderItemByStatus(EnumSet.allOf(OrderItem.OrderItemStatus.class))) {
+            orderItem.setStatus(OrderItem.OrderItemStatus.REMOVED);
+            orderItem.setDispatch(null);
         }
 
         orderHome.getInstance().setResReceived(false);
@@ -92,7 +97,7 @@ public class OrderCancel {
 
         initProcessInstanceHome();
 
-        if ("persisted".equals(orderHome.persist())) {
+        if ("updated".equals(orderHome.update())) {
             processInstanceHome.stop();
         }
